Move role survey and fact quiz components to own directories
