+++ frontend/src/components/factQuiz/CorrectAnswersInfo.jsx
