+++ frontend/src/components/factQuiz/FactInfoBox.jsx
