+++ frontend/src/components/roleSurvey/RoleAccordion.jsx
@@ -5,9 +5,9 @@ import {
     Typography,
 } from '@mui/material'
 import ExpandMoreIcon from '@mui/icons-material/ExpandMore'
-import { theme } from '../theme'
+import { theme } from '../../theme'
 import RoleImage from './RoleImage'
-import { RolePropTypes } from '../types'
+import { RolePropTypes } from '../../types'
 
 const RoleAccordion = ({ role }) => (
     <Accordion>
