+++ frontend/src/components/roleSurvey/SummaryDoughnut.jsx
