+++ frontend/src/pages/RoleSurveyQuestionPage.jsx
@@ -2,8 +2,8 @@ import { useEffect, useState } from 'react'
 import { useParams, useNavigate } from 'react-router-dom'
 import useSWR from 'swr'
 import { Stack, Card, CardContent, LinearProgress } from '@mui/material'
-import SurveyQuestionCard from '../components/SurveyQuestionCard'
-import RoleQuestionButtons from '../components/RoleQuestionButtons'
+import SurveyQuestionCard from '../components/roleSurvey/SurveyQuestionCard'
+import RoleQuestionButtons from '../components/roleSurvey/RoleQuestionButtons'
 import { useTitle } from '../hooks/useTitle'
 import { useSwipeable } from 'react-swipeable'
